backend/pyproject.toml
[tool.poetry]
name = "Parking Management System"
version = "0.1.0"
description = ""
authors = ["Samuel Agyei <88740526+SterotECH@users.noreply.github.com>"]
license = "Apache License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-unfold = "^0.6.5"
django-filter = "^23.2"
django-markdown = "^0.8.4"
python-dotenv = "^1.0.0"
django-cors-headers = "^4.1.0"
django-mail-templated = "^2.6.5"
daphne = "^4.0.0"
channels = "^4.0.0"
psycopg2 = "^2.9.6"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
